Add files via upload
This file contains the work completed for Task 2 of my internship using the Titanic dataset. The analysis includes:

Dropdown filters for passenger class and embarkation point using slicers.
A filter for gender selection.
Visualization of survival by gender on a pie chart.
Visualization of survival by passenger class on a donut chart.
A line and clustered column chart for passengers by embarkation point.